Expanded to create TODOs per users added support for JWT auths added admin for users with role admin added users endpoint for profile and password changing

+++ TodoApp/routers/todos.py
@@ -5,6 +5,7 @@ from sqlalchemy.orm import Session
 from models import Todos
 from database import SessionLocal
 from starlette import status
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -17,6 +18,8 @@ def get_db():
 
 db_dependency = Annotated[Session, Depends(get_db)]
 
+user_dependency = Annotated[dict, Depends(get_current_user)]
+
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
     description: str = Field(min_length=3, max_length=100)
@@ -24,28 +27,41 @@ class TodoRequest(BaseModel):
     complete: bool
 
 @router.get("/")
-async def read_all(db: db_dependency, status_code=status.HTTP_200_OK):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency, db: db_dependency, status_code=status.HTTP_200_OK):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    return db.query(Todos).filter(Todos.owner_id == user.get("id")).all()
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is not None:
         return todo_model
+
     raise HTTPException(status_code=404, detail="Todo not found.")
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+async def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    todo_model = Todos(**todo_request.model_dump(), owner_id=user.get("id"))
     
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency,
+async def update_todo(user: user_dependency, db: db_dependency,
                       todo_id: int,
                       todo_request: TodoRequest):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
@@ -59,11 +75,14 @@ async def update_todo(db: db_dependency,
     db.commit()
 
 @router.delete("/todo/{todo_id}")
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).delete()
     db.commit()
